CustomExceptionPractice/src/main/java/org/skillovilla/level2/BankAccount.java
package org.skillovilla.level2;

public class BankAccount {
    private String accountNumber;
    private double balance;

    public BankAccount(String accountNumber, double balance) {
        this.accountNumber = accountNumber;
        this.balance = balance;
    }

    public void withdraw(double amount) throws InvalidTransactionException {
        if (amount > balance) {
            throw new InvalidTransactionException("Insufficient funds");
        }
        balance -= amount;
    }

    public void deposit(double amount) throws InvalidTransactionException {
        if (amount <= 0) {
            throw new InvalidTransactionException("Invalid amount");
        }
        balance += amount;
    }

    public double getBalance() {
        return balance;
    }

    public String getAccountNumber() {
        return accountNumber;
    }

}
